requirements.txt
torch
torchvision
torchaudio
transformers==4.44.2
streamlit
pillow
datasets
evaluate
